12部署在服务器上/demo12a/demo12a/requirements.txt
asgiref==3.2.7
Django==2.1.4
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.1
